refactor: move all animation to own module
