+++ src/app/app.component.css
@@ -32,101 +32,3 @@ app-toast {
   z-index: 99;
   position: fixed;
 }
-
-/* Optimizaciones de performance para animaciones */
-.liquid-blob {
-  filter: blur(40px);
-  backface-visibility: hidden;
-  transform-style: preserve-3d;
-  /* Forzar compositing layer para mejor performance */
-  will-change: transform, border-radius;
-  /* Usar GPU acelerada */
-  transform: translateZ(0);
-}
-
-.central-blob {
-  filter: blur(60px);
-  backface-visibility: hidden;
-  transform-style: preserve-3d;
-  will-change: transform;
-  transform: translateZ(0);
-}
-
-/* Optimización para elementos flotantes */
-.floating-orb {
-  backface-visibility: hidden;
-  transform-style: preserve-3d;
-  will-change: transform;
-  transform: translateZ(0);
-}
-
-/* Contenedor principal optimizado */
-.backface-hidden {
-  backface-visibility: hidden;
-  transform-style: preserve-3d;
-}
-
-/* Optimizaciones para el elemento follower */
-.follower-element {
-  backface-visibility: hidden;
-  transform-style: preserve-3d;
-  /* Reducir repaints usando contain */
-  contain: layout style paint;
-  z-index: 110;
-}
-
-/* Mejoras de performance para contenedores aplicadas arriba en .site-container */
-
-/*
- * Accessibility: Comprehensive support for prefers-reduced-motion
- * These CSS rules complement the Angular MediaMatcher logic for better fallback support
- */
-@media (prefers-reduced-motion: reduce) {
-  /* Disable all CSS animations and transitions */
-  .liquid-blob,
-  .central-blob,
-  .floating-orb {
-    animation: none !important;
-    transition: none !important;
-    /* Disable transform animations but keep static positioning */
-    transform: translateZ(0) !important;
-  }
-
-  /* Reduce will-change to prevent unnecessary GPU layers for reduced motion */
-  .liquid-blob,
-  .central-blob,
-  .floating-orb {
-    will-change: auto !important;
-  }
-
-  /* Remove blur effects which can be motion-inducing */
-  .liquid-blob {
-    filter: blur(20px) !important; /* Reduced from 40px */
-  }
-
-  .central-blob {
-    filter: blur(30px) !important; /* Reduced from 60px */
-  }
-
-  /* Disable any parallax or scroll-triggered animations */
-  * {
-    animation-duration: 0.01ms !important;
-    animation-iteration-count: 1 !important;
-    transition-duration: 0.01ms !important;
-    scroll-behavior: auto !important;
-  }
-
-  /* Remove any potential motion from background gradients */
-  [style*="background"] {
-    background-attachment: scroll !important;
-  }
-}
-
-/* Optimización para pantallas de alta densidad */
-@media (-webkit-min-device-pixel-ratio: 2), (min-resolution: 192dpi) {
-  .liquid-blob,
-  .central-blob {
-    /* Reducir blur en pantallas HiDPI para mejor performance */
-    filter: blur(35px);
-  }
-}
